moving types to domain/types
