+++ Clients/src/domain/types/File.ts
