+++ Clients/src/domain/types/Framework.ts
