+++ Clients/src/domain/types/VendorRisk.ts
